refactor: Refactor HTML map export
